Massive updat to IJ Aspects (#515)
* Fix target platform to not bring in transitives

The transitives cause issues in self-hosted Eclipse with duplicate
bundles. We should be a good citizen and not contribute duplicate
bundles to the target platform.

* Add duplicate Guava because

See https://github.com/protocolbuffers/protobuf/issues/21173

* Launch configs

* Remove singleton pattern from info cache class

The lifetime of the cache is tightly coupled to the model manager. Thus,
it makes sense to move this into the model manager class, which forces
better design.

* Fix a problem where errors are not propagated properly to the UI

* Massive update to IJ Aspects

The IJ Aspects progressed quite a bit. A new templating system was
introduced to create aspects with different behavior depending on the
Bazel version.

Additional the design was modified so that aspects are no longer added
as a repository but become part of the workspace, i.e. they are copied
into the workspace during sync. This seems to work better with Bazel
caching (according to the IJ Aspects commit history).

When importing aspects we no longer consume them as a repository.
Instead we build them directly in the IJ repository and then copy &
extract into our zip archive. The zip archive is later extracted into a
temp location for better working with multiple Bazel workspaces.

* Add macOS cache location